+++ src/test/Stack/NixSpec.hs
@@ -12,8 +12,10 @@ import Prelude -- to remove the warning about Data.Monoid being redundant on GHC
 import Stack.Config
 import Stack.Config.Nix
 import Stack.Types.Config
+import Stack.Types.Compiler
 import Stack.Types.Nix
 import Stack.Types.StackT
+import Stack.Types.Version
 import System.Directory
 import System.Environment
 import System.IO.Temp (withSystemTempDirectory)
@@ -57,8 +59,10 @@ spec = beforeAll setup $ afterAll teardown $ do
       writeFile (toFilePath stackDotYaml) sampleConfig
       lc <- loadConfig' manager
       (nixEnable $ configNix $ lcConfig lc) `shouldBe` True
-    it "sees that the only package asked for is glpk and adds GHC from nixpkgs mirror of LTS resolver" $ \T{..} -> inTempDir $ do
-      writeFile (toFilePath stackDotYaml) sampleConfig
-      lc <- loadConfig' manager
-      (nixPackages $ configNix $ lcConfig lc) `shouldBe` ["glpk"]
-      nixCompiler (lcConfig lc) Nothing Nothing `shouldBe` "haskell.packages.lts-2_10.ghc"
+    it "sees that the only package asked for is glpk and asks for the correct GHC derivation" $
+      \T{..} -> inTempDir $ do
+        writeFile (toFilePath stackDotYaml) sampleConfig
+        lc <- loadConfig' manager
+        (nixPackages $ configNix $ lcConfig lc) `shouldBe` ["glpk"]
+        v <- parseVersion "7.10.3"
+        nixCompiler (GhcVersion v) `shouldBe` "haskell.compiler.ghc7103"